Add location to record objects
The web app expects record objects returned by this API to include a
location, so this commit updates the API accordingly.
