+++ packages/api/src/common/models.ts
@@ -6,3 +6,17 @@ export enum AccessRole {
 	Contributor = "access.role.contributor",
 	Curator = "access.role.curator",
 }
+
+export interface Location {
+	id: string;
+	streetNumber?: string;
+	streetName?: string;
+	locality?: string;
+	county?: string;
+	state?: string;
+	latitude?: number;
+	longitude?: number;
+	country?: string;
+	countryCode?: string;
+	displayName?: string;
+}
