+++ packages/api/src/record/fixtures/create_complete_test_record.sql
@@ -22,7 +22,8 @@ record (
   type,
   createddt,
   updateddt,
-  alttext
+  alttext,
+  locnid
 )
 VALUES
 (
@@ -48,5 +49,6 @@ VALUES
   'type.record.image',
   '2023-06-21T00:00:00.000Z',
   '2023-06-21T00:00:00.000Z',
-  'An image'
+  'An image',
+  1
 );
