+++ packages/api/src/record/fixtures/create_test_locations.sql
@@ -1,0 +1,27 @@
+INSERT INTO locn (
+  locnid,
+  streetnumber,
+  streetname,
+  locality,
+  admintwoname,
+  latitude,
+  longitude,
+  country,
+  countrycode,
+  displayname,
+  status,
+  type
+) VALUES (
+  1,
+  '55',
+  'Rue Plumet',
+  'Paris',
+  'Ile-de-France',
+  48.838608548520966,
+  2.3069214988665303,
+  'France',
+  'FR',
+  'Jean Valjean''s House',
+  'status.generic.ok',
+  'type.account.standard'
+);
